requirements.txt
flask>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
